move SensorValue into sensor package
