+++ src/main/java/org/chabala/brick/controllab/sensor/SensorValueImpl.java
@@ -16,7 +16,7 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with brick-control-lab.  If not, see http://www.gnu.org/licenses/.
  */
-package org.chabala.brick.controllab;
+package org.chabala.brick.controllab.sensor;
 
 /**
  * {@inheritDoc}.
